requirements.txt
scikit-learn==1.4.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
pandas==2.2.1
matplotlib-inline==0.1.7
scipy==1.14.1
panda==0.3.1
tqdm==4.66.1